Changed the ExerciseActivity view to display UserExercise details rather than activity types (this is a temporary feature during development).
Added getDatePlus method to DataBaseInitializer class to add minutes on to a Calendar object.

Added findByID method to ActivityDAO to return an activity for an id.

Added onTimeSet method to AddNewExerciseActivity view this sets the hour and minute aspects to the startdate/time Calendar object.  The onDateSet method has bee changed to now show the time picker dialog when the date has been set.

Added listeners for the two duration fields.  This allows variables to be set on the afterTextChanged events for both hours and minutes.  Method also added to be used by button click to submit the duration of the activity.  Duration in entered in hours and minutes though is converted to minutes for calculation and storage.

TimeSelectFragment added to hold the time picker dialog.

Combined the start date and time fields in the UserExercise entity to a single startDateAndTime field.