Renamed activityDAO to activityModel using IDE rename.
Added ExerciseListViewModel.java as a ViewModel for ExerciseActivity which is th main list of all exercise for that user.  At present this ViewModel and the others only handles making db requests and working with the entities to remove this work from the view.  Intended to move more done in views (click events etc) to ViewModels.